requirements.txt
Flask==1.1.2
numpy==1.18.1
pandas ==0.25.3
flask-feather==0.1.0
sklearn ==0.0
scikit-learn==0.23.1
gunicorn==20.0.4
feather-format==0.4.1
